feat: make a custom hook to add animation props
